requirements.txt
torch
torchvision
pyyaml
cv2
pycocotools